Moved request/response structs into router
