+++ server/src/routes/routes_todo.rs
@@ -5,16 +5,50 @@ use axum::http::StatusCode;
 use axum::routing::{get, patch, post, put};
 use axum::{Json, Router};
 use chrono::DateTime;
+use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use uuid::Uuid;
 
 use crate::app_state::AppState;
 use crate::controllers::todo_controller::ToDoController;
 use crate::errors::to_do_error::ToDoError;
-use crate::models::todo::{
-    RequestCreateToDoItem, RequestUpdateToDoItemCompleted, RequestUpdateToDoItemRank,
-    ResponseGetToDoByDate,
-};
+use crate::models::todo::ToDoItemWithRank;
+
+// **********************************************
+// * REQUEST/RESPONSE TYPES
+// **********************************************
+
+#[typeshare::typeshare]
+#[derive(Debug, Deserialize)]
+pub struct RequestCreateToDoItem {
+    pub title: String,
+    pub description: String,     // Make this markdown later
+    pub dates: Vec<Vec<String>>, // The inner vector should just be of length 2 for every entry. First string is the ISO date string and the second is the rank
+}
+
+#[typeshare::typeshare]
+#[derive(Debug, Deserialize)]
+pub struct RequestUpdateToDoItemRank {
+    pub iso_string: String,
+    pub item_id: Uuid,
+    pub new_rank: String,
+}
+
+#[typeshare::typeshare]
+#[derive(Debug, Deserialize)]
+pub struct RequestUpdateToDoItemCompleted {
+    pub completed: bool,
+}
+
+#[typeshare::typeshare]
+#[derive(Debug, Serialize)]
+pub struct ResponseGetToDoByDate {
+    pub items: Vec<ToDoItemWithRank>,
+}
+
+// **********************************************
+// * ROUTES
+// **********************************************
 
 pub fn routes(app_state: AppState) -> Router {
     Router::new()
